add explicit_memset and fix explicit_bzero with it
Clang, GCC and other compilers special-case standard C functions like
memset. Calls to memset will be optimized out if nothing uses the
result, defeating usage as an information sanitization mechanism.

OpenBSD provides explicit_bzero to work around this but Android simply
defines it as memset so nothing prevents it from being optimized away.

The explicit_memset function comes from NetBSD and can be used as a
drop-in replacement for memset with the guarantee that it won't be
optimized out.

This implementation uses a memory read constraint via empty inline
assembly rather than something that may be broken via link-time
optimization in the future.

Change-Id: Ia098f3ea0f670361f82b21083208e1641a714848